backend/requirements.txt
dj-database-url==0.5.0
dj-rest-auth==2.1.11
Django==4.0.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
mypy-extensions==0.4.3
oauthlib==3.1.1
PyJWT==2.3.0
regex==2021.11.10
psycopg2-binary==2.9.3
django-cors-headers==3.13.0
sqlparse==0.4.2
django-environ==0.9.0
pytz==2021.3
